storybook: rename config files to mjs
